+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - New listener for Jetpack.
 - New listener for WP Security Audit Log.
 - New listener for W3 Total Cache.
+- New listener for WP Super Cache.
 - New logger to send logs to Solawinds Loggly.
 - New logger to send logs to Logentries / insightOps.
 - New "Content" box in event viewer (for WordPress events logs) to display detailed error code and message.